Methods

Index

AstroPlot.axis_cartesian

AstroPlot.axisid

AstroPlot.densitycurve

AstroPlot.distribution

AstroPlot.force

AstroPlot.lagrange_radii

AstroPlot.pack_xy

AstroPlot.plot_lagrangeradii

AstroPlot.plot_lagrangeradii!

AstroPlot.plot_lagrangeradii90

AstroPlot.plot_lagrangeradii90!

AstroPlot.plot_makie

AstroPlot.plot_makie!

AstroPlot.plot_momentum

AstroPlot.plot_momentum_angular

AstroPlot.plot_peano

AstroPlot.plot_radii

AstroPlot.plot_radii!

AstroPlot.plot_scaleradius

AstroPlot.plot_slice

AstroPlot.png2video

AstroPlot.pos_from_center

AstroPlot.projection_density

AstroPlot.radialforce

AstroPlot.radialvel

AstroPlot.rotationcurve

AstroPlot.rotvel

AstroPlot.sortarrays!

AstroPlot.unicode_density

AstroPlot.unicode_force

AstroPlot.unicode_projection

AstroPlot.unicode_projection_density

AstroPlot.unicode_radialforce

AstroPlot.unicode_rotationcurve

AstroPlot.unicode_scatter

AstroPlot.unicode_slice

MosaicViews.mosaicview

UnicodePlots

AstroPlot.unicode_scatter — Function

function unicode_scatter(data::Array{T, N}, u::Union{Nothing, Unitful.FreeUnits} = u"kpc"; kw...) where T<:AbstractPoint where N
function unicode_scatter(data::Array{T, N}, u::Union{Nothing, Unitful.FreeUnits} = u"kpc"; kw...) where T<:AbstractParticle where N
function unicode_scatter(data::StructArray{T,N,NT,Tu}, u::Union{Nothing, Unitful.FreeUnits} = u"kpc"; kw...) where T<:AbstractParticle where N where NT where Tu
function unicode_scatter(data::Union{Array{T,N}, StructArray{T,N,NT,Tu}}, collection::Collection, u::Union{Nothing, Unitful.FreeUnits} = u"kpc"; kw...) where T<:AbstractParticle where N where NT where Tu

Scatter plot of points or particles in REPL

collection: filter the type of particles

Keywords

xaxis: Symbol to plot on x-axis. Default is :x

yaxis: Symbol to plot on y-axis. Defualt is :y

xlabel: label of x-axis

ylabel: label of y-axis

Examples

julia> unicode_scatter(randn_pvector(100))

AstroPlot.unicode_density — Function

function unicode_density(data, units = uAstro; kw...)

Plot radial mass density curve of spherically symmetric system in REPL

Keywords

xlabel: label of x-axis

ylabel: label of y-axis

title: title line of the figure

timestamp: print time on figure title

AstroPlot.unicode_force — Function

unicode_force(data, units = uAstro; kw...)

Plot acceleration magnitude distribution of spherically symmetric system in REPL

Keywords

xlabel: label of x-axis

ylabel: label of y-axis

title: title line of the figure

timestamp: print time on figure title

AstroPlot.unicode_radialforce — Function

unicode_radialforce(data, units = uAstro; kw...)

Plot radial acceleration magnitude relative to center in REPL

Keywords

xlabel: label of x-axis

ylabel: label of y-axis

title: title line of the figure

timestamp: print time on figure title

Missing docstring.

Missing docstring for unicode_radialpotential. Check Documenter's build log for details.

AstroPlot.unicode_rotationcurve — Function

function unicode_rotationcurve(data, units = uAstro; kw...)

Plot rotation curve in REPL

Keywords

rmhead: remove the first rmhead elements of result array. Useful to avoid improper data points

rmtail: remove the last rmtail elements of result array. Useful to avoid improper data points

savelog::Bool: if true, save processed data in csv. The name of log file depends on analysis function

savefolder: set the directory to save log file

section::Int64: number of bins. Default is floor(Int64, length(data)^0.5833333333333334)

AstroPlot.unicode_projection — Function

function unicode_projection(ρ; kw...)

Plot 2D projection sum over one dimension.

Keywords

xaxis: Symbol to plot on x-axis. Default is :x

yaxis: Symbol to plot on y-axis. Defualt is :y

xlabel: label of x-axis

ylabel: label of y-axis

title: title line of the figure

colormap = :inferno: color map of heatmap plot

AstroPlot.unicode_projection_density — Function

function unicode_projection_density(mesh::MeshCartesianStatic; kw...)

Plot 2D projection sum of mesh density over one dimension.

Keywords

xaxis: Symbol to plot on x-axis. Default is :x

yaxis: Symbol to plot on y-axis. Defualt is :y

xlabel: label of x-axis

ylabel: label of y-axis

title: title line of the figure

colormap = :inferno: color map of heatmap plot

AstroPlot.unicode_slice — Function

function unicode_slice(data::AbstractArray{T,3}, n::Int; kw...)

Plot 2D slice of a 3D array. Vector plot is not supported in unicode mode.

Keywords

xaxis: Symbol to plot on x-axis. Default is :x

yaxis: Symbol to plot on y-axis. Defualt is :y

xlabel: label of x-axis

ylabel: label of y-axis

colormap = :inferno: color map of heatmap plot

function unicode_slice(mesh::MeshCartesianStatic, symbol::Symbol, n::Int;

Plot 2D slice of 3D data symbol in mesh.

Keywords

xaxis: Symbol to plot on x-axis. Default is :x

yaxis: Symbol to plot on y-axis. Defualt is :y

xlabel: label of x-axis

ylabel: label of y-axis

colormap = :inferno: color map of heatmap plot

GLMakie

AstroPlot.plot_makie — Function

plot_makie(data::Array{T,1}, u::Union{Nothing, Unitful.FreeUnits} = u"kpc"; kw...) where T<:PVector
plot_makie(data::Array{T,1}, u::Union{Nothing, Unitful.FreeUnits} = u"kpc"; kw...) where T<:AbstractParticle3D
plot_makie(data::StructArray, u::Union{Nothing, Unitful.FreeUnits} = u"kpc"; kw...)
plot_makie(data::Union{Array{T,N}, StructArray{T,N,NT,Tu}}, collection::Collection, u::Union{Nothing, Unitful.FreeUnits} = u"kpc"; kw...) where T<:AbstractParticle3D where N where NT where Tu

Plot scatter data points (in interactive mode)

collection: filter the type of particles

Examples

d = randn_pvector(50)
plot_makie(d, nothing)

d = randn_pvector(50, u"km")
plot_makie(d, u"m")

plot_makie(nodes::Array{T,N}; kw...) where T<:PhysicalTrees.OctreeNode where N

Plot tree nodes in wireframe mode

AstroPlot.plot_makie! — Function

plot_makie!(scene, data::Array{T,1}, where T<:AbstractParticle3D where N where NT where Tu u"kpc"; kw...) where T<:PVector
plot_makie!(scene, data::Array{T,1}, where T<:AbstractParticle3D where N where NT where Tu u"kpc"; kw...) where T<:AbstractParticle3D
plot_makie!(scene, data::StructArray, where T<:AbstractParticle3D where N where NT where Tu u"kpc"; kw...)
plot_makie!(scene, data::Union{Array{T,N}, StructArray{T,N,NT,Tu}}, collection::Collection, u::Union{Nothing, Unitful.FreeUnits} = u"kpc"; kw...) where T<:AbstractParticle3D where N where NT where Tu

Plot scatter data points (in interactive mode)

collection: filter the type of particles

Examples

d = randn_pvector(50)
plot_makie!(scene, d, nothing)

d = randn_pvector(50, u"km")
plot_makie!(scene, d, u"m")

plot_makie!(axis, nodes::Array{T,N}; kw...) where T<:PhysicalTrees.OctreeNode where N

Plot tree nodes in wireframe mode

AstroPlot.plot_peano — Function

plot_peano(bits::Int = 1)

Interactively plot Hilber-Peano curve with Makie. It is recommanded that bits <= 6 to avoid lagging problems.

CairoMakie

Missing docstring.

Missing docstring for projection. Check Documenter's build log for details.

AstroPlot.projection_density — Function

function projection_density(mesh::MeshCartesianStatic; kw...)

Plot 2D projection sum of mesh density over one dimension.

Keywords

aspect_ratio: aspect ratio of axes. Default is 1.0 to avoid stretching. Pass to Makie as AxisAspect(aspect_ratio)

resolution: figure size

xaxis: Symbol to plot on x-axis. Default is :x

yaxis: Symbol to plot on y-axis. Defualt is :y

xlabel: label of x-axis

ylabel: label of y-axis

title: title line of the figure

xlims: set x-limit of the plot. Default is nothing

ylims: set y-limit of the plot. Default is nothing

AstroPlot.plot_slice — Function

function plot_slice(mesh::MeshCartesianStatic, symbol::Symbol, n::Int;

Plot 2D slice of 3D data symbol in mesh.

Keywords

resolution: figure size

xaxis: Symbol to plot on x-axis. Default is :x

yaxis: Symbol to plot on y-axis. Defualt is :y

xlabel: label of x-axis

ylabel: label of y-axis

title: title line of the figure

xlims: set x-limit of the plot. Default is nothing

ylims: set y-limit of the plot. Default is nothing

aspect_ratio: aspect ratio of axes. Default is 1.0 to avoid stretching. Pass to Makie as AxisAspect(aspect_ratio)

Video

AstroPlot.png2video — Function

function png2video(dir::String, filenamebase::String, suffix::String, outfile::String; kw...)

Convert all figure end with suffix to video. Filenames are organized as filenamebase0000suffix. For png files like pos_0000.png, set filenamebase = "pos_" and suffix = ".png"

It basically calls VideoIO.open_video_out

Keywords

encoder_options: keywords supported by FFMPEG. Default value (crf=23, preset="medium") is suitable for most cases

framerate: fps (frame per second). Default 24

codec_name: name of video encoder. Default nothing. Try libx264, libx265, libx264rgb and more options in FFMPEG

Examples

julia> png2video("output", "pos_", ".png", "test.mp4")

MosaicViews.mosaicview — Function

function mosaicview(folder::String, filenamebase::String, Counts::Array, suffix::String; kw...)

Arguments

folder: directory holding snapshots

filenamebase, Counts, suffix: snapshots are commonly named as snapshot_0000.gadget2, in this way filenamebase = "snapshot_", suffix = ".gadget2. Counts is an array to choose snapshots, and it is printed to formatted string (controled by keyword formatstring) in for loops.

Examples

julia> mosaicview("output", "pos_", collect(0:10:490), ".png"; fillvalue=0.5, npad=5, ncol=10, rowmajor=true)

function mosaicview(folder::String, filenamebase::String, suffix::String; kw...)

Plot all files matching filenamebase and suffix in folder to mosaic view.

Tools

Analysis

AstroPlot.rotationcurve — Function

function rotationcurve(data, units = uAstro; kw...)

Return Tuple(Rmean, Vmean, Rstd, Vstd), where mean is mean value, std means standard deviation.

Keywords

rmhead: remove the first rmhead elements of result array. Useful to avoid improper data points

rmtail: remove the last rmtail elements of result array. Useful to avoid improper data points

savelog::Bool: if true, save processed data in csv. The name of log file depends on analysis function

savefolder: set the directory to save log file

section::Int64: number of bins. Default is floor(Int64, length(data)^0.5833333333333334)

AstroPlot.densitycurve — Function

function densitycurve(data, units = uAstro; kw...)

Compute radial mass density curve of spherically symmetric system

Keywords

savelog::Bool: if true, save processed data in csv. The name of log file depends on analysis function

savefolder: set the directory to save log file

AstroPlot.rotvel — Function

rotvel(vel::PVector, pos::PVector)

Rotational velocity magnitude at pos relative to origin.

AstroPlot.radialvel — Function

radialvel(vel::PVector, pos::PVector)

Radial velocity magnitude at pos relative to origin.

AstroPlot.force — Function

force(data, units = uAstro; kw...)

Compute acceleration magnitude distribution of spherically symmetric system

Keywords

savelog::Bool: if true, save processed data in csv. The name of log file depends on analysis function

savefolder: set the directory to save log file

AstroPlot.radialforce — Function

radialforce(a0, acc, p0, pos)

Compute radial acceleration magnitude relative to center p0

radialforce(data, units = uAstro; kw...)

Compute radial acceleration magnitude relative to center

Keywords

savelog::Bool: if true, save processed data in csv. The name of log file depends on analysis function

savefolder: set the directory to save log file

Missing docstring.

Missing docstring for radialpotential. Check Documenter's build log for details.

AstroPlot.distribution — Function

function distribution(x::Array, y::Array; kw...)

Compute how quantity y is distributed along x. For example, rotation velocity along radius. Return Tuple(xmean, ymean, xstd, ystd), where mean is mean value, std means standard deviation.

Keywords

section::Int64: number of bins. Default is floor(Int64, length(data)^0.5833333333333334)

rmhead: remove the first rmhead elements of result array. Useful to avoid improper data points

rmtail: remove the last rmtail elements of result array. Useful to avoid improper data points

AstroPlot.pos_from_center — Function

function pos_from_center(particles, u = u"kpc")
function pos_from_center(pos::Array{T,1}, u = u"kpc") where T <: AbstractPoint

Return a unitless array of positions relative to the center.

AstroPlot.lagrange_radii — Function

function lagrange_radii(data, u = u"kpc")

Return a Tuple of scale radius and Lagrange radii. Designed for spherically symmetric systems.

Data processing

AstroPlot.pack_xy — Function

function pack_xy(data, u = nothing; kw...)

Return Tuple (x,y) substracted from points or positions of particles in data. Useful to prepare a plot.

Keywords

xaxis: Symbol to plot on x-axis. Default is :x

yaxis: Symbol to plot on y-axis. Defualt is :y

AstroPlot.sortarrays! — Function

function sortarrays!(by::Array, another::Array; kw...)

Sort two arrays at the same time. kw is passed to Base.sort!.

Examples

julia> a = [1,3,2]
julia> b = ["a", "b", "c"]
julia> sortarrays!(a,b)
julia> a
3-element Vector{Int64}:
 1
 2
 3

julia> b
3-element Vector{String}:
 "a"
 "c"
 "b"

Missing docstring.

Missing docstring for pos_from_center. Check Documenter's build log for details.

Meshes

AstroPlot.axisid — Function

function axisid(axis::Symbol)

Convert axis to Int

return 1 if axis == :x

return 2 if axis == :y

return 3 if axis == :z

AstroPlot.axis_cartesian — Function

function axis_cartesian(pos::StructArray, axis::Symbol)
function axis_cartesian(mesh::MeshCartesianStatic, axis::Symbol)

Return 1D axis points of 3D Cartesian positions.

Missing docstring.

Missing docstring for slice3d. Check Documenter's build log for details.

energy evolution

plot_energy
plot_energy_kinetic
plot_energy_potential
plot_energy_delta
plot_energy_delta!
kinetic_energy
sum_kinetic
sum_potential

momentum evolution

AstroPlot.plot_momentum — Function

plot_momentum(df::DataFrame; kw...)

Plot total momentum in df which contains columns named time and momentum

Keywords

resolution: figure size

xaxis: Symbol to plot on x-axis. Default is :x

yaxis: Symbol to plot on y-axis. Defualt is :y

xlabel: label of x-axis

ylabel: label of y-axis

title: title line of the figure

xlims: set x-limit of the plot. Default is nothing

ylims: set y-limit of the plot. Default is nothing

plot_momentum(datafile::String; kw...)

Plot total momentum in datafile which is a CSV file containing columns named time and momentum

Keywords

resolution: figure size

xaxis: Symbol to plot on x-axis. Default is :x

yaxis: Symbol to plot on y-axis. Defualt is :y

xlabel: label of x-axis

ylabel: label of y-axis

title: title line of the figure

xlims: set x-limit of the plot. Default is nothing

ylims: set y-limit of the plot. Default is nothing

plot_momentum(folder::String, filenamebase::String, Counts::Vector{Int64}, suffix::String, FileType::AbstractOutputType, units = uAstro, fileunits = uGadget2; kw...)

Plot total momentum and total angular momentum of particles in each snapshot.

Arguments

folder: directory holding snapshots

filenamebase, Counts, suffix: snapshots are commonly named as snapshot_0000.gadget2, in this way filenamebase = "snapshot_", suffix = ".gadget2. Counts is an array to choose snapshots, and it is printed to formatted string (controled by keyword formatstring) in for loops.

times: set time label of each snapshot

FileType: Trait argument to dispatch on different snapshot format. It is to be set manually because some output types cannot be deduced automatically

u = u"kpc": length unit. Set u as nothing in unitless cases.

Keywords

formatstring: formatted string to control snapshot index. Default is "%04d"

resolution: figure size

xaxis: Symbol to plot on x-axis. Default is :x

yaxis: Symbol to plot on y-axis. Defualt is :y

xlabel: label of x-axis

ylabel: label of y-axis

title: title line of the figure

xlims: set x-limit of the plot. Default is nothing

ylims: set y-limit of the plot. Default is nothing

savelog::Bool: if true, save processed data in csv. The name of log file depends on analysis function

savefolder: set the directory to save log file

AstroPlot.plot_momentum_angular — Function

plotmomentumangular(df::DataFrame; kw...)

Plot total angular momentum in df which contains columns named time and angularmomentum

Keywords

resolution: figure size

xaxis: Symbol to plot on x-axis. Default is :x

yaxis: Symbol to plot on y-axis. Defualt is :y

xlabel: label of x-axis

ylabel: label of y-axis

title: title line of the figure

xlims: set x-limit of the plot. Default is nothing

ylims: set y-limit of the plot. Default is nothing

plot_momentum_angular(datafile::String; kw...)

Plot total angular momentum in datafile which is a CSV file containing columns named time and angularmomentum

Keywords

resolution: figure size

xaxis: Symbol to plot on x-axis. Default is :x

yaxis: Symbol to plot on y-axis. Defualt is :y

xlabel: label of x-axis

ylabel: label of y-axis

title: title line of the figure

xlims: set x-limit of the plot. Default is nothing

ylims: set y-limit of the plot. Default is nothing

radii evolution

AstroPlot.plot_scaleradius — Function

function plot_scaleradius(df::DataFrame, uTime::Units, uLength::Units; kw...)

Plot evolution of scale radius by time. df contains columns named Time and ScaleRadius

Return a Tuple of scene and layout

Keywords

xlabel: label of x-axis

ylabel: label of y-axis

title: title line of the figure

resolution: figure size

AstroPlot.plot_lagrangeradii — Function

function plot_lagrangeradii(df::DataFrame, uTime::Units, uLength::Units; kw...)

Plot evolution of Lagrange radii by time. df contains columns named Time and L10, L20, ..., L100

Return a Tuple of scene and layout

Keywords

title: title line of the figure

xaxis: Symbol to plot on x-axis. Default is :x

yaxis: Symbol to plot on y-axis. Defualt is :y

xlabel: label of x-axis

ylabel: label of y-axis

AstroPlot.plot_lagrangeradii! — Function

function plot_lagrangeradii!(scene, ax, layout, df::DataFrame; kw...)

Plot evolution of Lagrange radii by time. df contains columns named Time and L10, L20, ..., L100

Return a Tuple of scenes and layouts of different line plots

Keywords

title: title line of the figure

xaxis: Symbol to plot on x-axis. Default is :x

yaxis: Symbol to plot on y-axis. Defualt is :y

xlabel: label of x-axis

ylabel: label of y-axis

AstroPlot.plot_lagrangeradii90 — Function

function plot_lagrangeradii90(df::DataFrame, uTime::Units, uLength::Units; kw...)

Plot evolution of Lagrange radii by time. df contains columns named Time and L10, L20, ..., L90 L100 is omitted, because in most cases, those escaping particles may over distort the figure.

Return a Tuple of scene and layout

Keywords

title: title line of the figure

xaxis: Symbol to plot on x-axis. Default is :x

yaxis: Symbol to plot on y-axis. Defualt is :y

xlabel: label of x-axis

ylabel: label of y-axis

AstroPlot.plot_lagrangeradii90! — Function

function plot_lagrangeradii90!(scene, ax, layout, df::DataFrame; kw...)

Plot evolution of Lagrange radii by time. df contains columns named Time and L10, L20, ..., L90 L100 is omitted, because in most cases, those escaping particles may over distort the figure.

Return a Tuple of scenes and layouts of different line plots

Keywords

title: title line of the figure

xaxis: Symbol to plot on x-axis. Default is :x

yaxis: Symbol to plot on y-axis. Defualt is :y

xlabel: label of x-axis

ylabel: label of y-axis

AstroPlot.plot_radii — Function

function plot_radii(folder::String, filenamebase::String, Counts::Array{Int64,1}, suffix::String, FileType::AbstractOutputType, units = uAstro; kw...)

Plot scale radius and Lagrange radii (up to 90% by default)

Return Tuple(ScaleScene, ScaleLayout, LagrangeScene, LagrangeLayout, df). df contains radii data

Arguments

folder: directory holding snapshots

filenamebase, Counts, suffix: snapshots are commonly named as snapshot_0000.gadget2, in this way filenamebase = "snapshot_", suffix = ".gadget2. Counts is an array to choose snapshots, and it is printed to formatted string (controled by keyword formatstring) in for loops.

times: set time label of each snapshot

FileType: Trait argument to dispatch on different snapshot format. It is to be set manually because some output types cannot be deduced automatically

u = u"kpc": length unit. Set u as nothing in unitless cases.

Keywords

formatstring: formatted string to control snapshot index. Default is "%04d"

Examples

julia> ScaleScene, ScaleLayout, LagrangeScene, LagrangeLayout, df = plotradii( joinpath(pathof(AstroPlot), "../../test/snapshots"), "snapshot", collect(0:20:200), ".gadget2", gadget2(), times = collect(0.0:0.01:0.1) * u"Gyr", title = "Radii plot")

AstroPlot.plot_radii! — Function

function plot_radii(AS, SL, AL, LL, folder::String, filenamebase::String, Counts::Array{Int64,1}, suffix::String, FileType::AbstractOutputType, units = uAstro; kw...)

Plot scale radius and Lagrange radii (up to 90% by default) Return radii data in Dict

Arguments

AS: Axis of scale radius scene

SL: Lagrange radii scene

AL: Axis of Lagrange radii scene

LL: layout of Lagrange radii scene

folder: directory holding snapshots

filenamebase, Counts, suffix: snapshots are commonly named as snapshot_0000.gadget2, in this way filenamebase = "snapshot_", suffix = ".gadget2. Counts is an array to choose snapshots, and it is printed to formatted string (controled by keyword formatstring) in for loops.

times: set time label of each snapshot

FileType: Trait argument to dispatch on different snapshot format. It is to be set manually because some output types cannot be deduced automatically

u = u"kpc": length unit. Set u as nothing in unitless cases.

Keywords

formatstring: formatted string to control snapshot index. Default is "%04d"

Examples

julia> ScaleScene, ScaleLayout, LagrangeScene, LagrangeLayout, df = plotradii( joinpath(pathof(AstroPlot), "../../test/snapshots"), "snapshot", collect(0:20:200), ".gadget2", gadget2(), times = collect(0.0:0.01:0.1) * u"Gyr", title = "Radii plot")

force distribution

plot_force
plot_force!

density distribution

Missing docstring.

Missing docstring for plot_density. Check Documenter's build log for details.

Missing docstring.

Missing docstring for plot_density!. Check Documenter's build log for details.